+++ crates/icu/src/macros/endpoints/root.rs
@@ -6,10 +6,10 @@ macro_rules! icu_endpoints_root {
         // modify app-level state
         // @todo eventually this will cascade down from an orchestrator canister
         #[::icu::cdk::update]
-        async fn icu_app(cmd: ::icu::memory::app::AppCommand) -> Result<(), ::icu::Error> {
+        async fn icu_app(cmd: ::icu::memory::state::AppCommand) -> Result<(), ::icu::Error> {
             $crate::auth_require_any!(::icu::auth::is_controller)?;
 
-            ::icu::memory::app::AppState::command(cmd)?;
+            ::icu::memory::state::AppState::command(cmd)?;
 
             let bundle = ::icu::ops::sync::state::StateBundle::root();
             ::icu::ops::sync::state::root_cascade(bundle).await?;
@@ -81,6 +81,11 @@ macro_rules! icu_endpoints_root {
         // POOL ENDPOINTS
         //
 
+        #[::icu::cdk::query]
+        fn icu_pool_list() -> ::icu::memory::root::CanisterPoolView {
+            $crate::memory::root::CanisterPool::export()
+        }
+
         #[update]
         async fn icu_pool_create_canister() -> Result<Principal, ::icu::Error> {
             $crate::auth_require_any!(::icu::auth::is_controller)?;
@@ -93,15 +98,6 @@ macro_rules! icu_endpoints_root {
             $crate::auth_require_any!(::icu::auth::is_controller)?;
 
             ::icu::ops::pool::move_canister_to_pool(pid).await
-        }
-
-        //
-        // MEMORY ENDPOINTS
-        //
-
-        #[::icu::cdk::query]
-        fn icu_canister_pool() -> ::icu::memory::root::CanisterPoolView {
-            $crate::memory::root::CanisterPool::export()
         }
 
         //