+++ crates/icu/src/macros/mod.rs
@@ -28,7 +28,7 @@ macro_rules! log {
     // Inner logic
     (@inner $level:expr, $fmt:expr, $($arg:tt)*) => {{
         let message = format!($fmt, $($arg)*);
-        let ty_raw = match $crate::memory::canister::CanisterState::get_view() {
+        let ty_raw = match $crate::memory::state::CanisterState::get_view() {
             Some(entry) => entry.ty.to_string(),
             None => "-".to_string(),
         };