+++ crates/icu/src/macros/start.rs
@@ -18,18 +18,19 @@ macro_rules! icu_start {
     ($canister_type:expr) => {
         #[::icu::cdk::init]
         fn init(
-            state: ::icu::memory::canister::CanisterStateData,
+            state: ::icu::memory::state::CanisterStateData,
             parents: Vec<::icu::memory::CanisterView>,
             args: Option<Vec<u8>>,
         ) {
+            ::icu::memory::state::CanisterState::import(state);
             ::icu::log!(::icu::Log::Info, "🏁 init: {}", $canister_type);
 
             // setup
-            ::icu::memory::canister::CanisterState::import(state);
             ::icu::memory::subnet::SubnetParents::import(parents);
             ::icu::memory::canister::CanisterRoot::set(::icu::cdk::api::msg_caller());
             __icu_shared_setup();
 
+            // timer - icu_install
             let _ = ::icu::cdk::timers::set_timer(::std::time::Duration::from_secs(0), move || {
                 ::icu::cdk::futures::spawn(icu_install(args));
             });
@@ -39,6 +40,7 @@ macro_rules! icu_start {
         fn post_upgrade() {
             __icu_shared_setup();
 
+            // timer - icu_upgrade
             let _ = ::icu::cdk::timers::set_timer(::std::time::Duration::from_secs(0), move || {
                 ::icu::cdk::futures::spawn(icu_upgrade());
             });
@@ -47,7 +49,9 @@ macro_rules! icu_start {
         #[allow(unexpected_cfgs)]
         fn __icu_shared_setup() {
             ::icu::__icu_load_config!();
-            ::icu::memory::cycles::CycleTracker::start();
+            ::icu::memory::registry::force_init_all_tls();
+            ::icu::memory::canister::CycleTracker::start();
+
             icu_setup();
         }
 
@@ -68,15 +72,14 @@ macro_rules! icu_start_root {
             );
             ::icu::log!(::icu::Log::Info, "🏁 init: root");
 
-            // register
+            // setup
             let entry =
                 ::icu::memory::subnet::SubnetRegistry::init_root(::icu::cdk::api::canister_self());
             ::icu::memory::canister::CanisterRoot::set(::icu::cdk::api::canister_self());
-            ::icu::memory::canister::CanisterState::set_view(entry.into());
+            ::icu::memory::state::CanisterState::set_view(entry.into());
+            __icu_root_shared_setup();
 
-            // setup
-            __icu_shared_setup();
-
+            // timer - icu_install
             let _ = ::icu::cdk::timers::set_timer(::std::time::Duration::from_secs(0), move || {
                 ::icu::cdk::futures::spawn(icu_install());
             });
@@ -84,19 +87,22 @@ macro_rules! icu_start_root {
 
         #[::icu::cdk::post_upgrade]
         fn post_upgrade() {
-            __icu_shared_setup();
+            __icu_root_shared_setup();
 
+            // timer - icu_upgrade
             let _ = ::icu::cdk::timers::set_timer(::std::time::Duration::from_secs(0), move || {
                 ::icu::cdk::futures::spawn(icu_upgrade());
             });
         }
 
         #[allow(unexpected_cfgs)]
-        fn __icu_shared_setup() {
+        fn __icu_root_shared_setup() {
             ::icu::__icu_load_config!();
+            ::icu::memory::registry::force_init_all_tls();
             ::icu::memory::root::CanisterPool::start();
-            ::icu::memory::cycles::CycleTracker::start();
+            ::icu::memory::canister::CycleTracker::start();
             ::icu::state::wasm::WasmRegistry::import(WASMS);
+
             icu_setup();
         }