+++ crates/icu/src/memory/state/app.rs
@@ -1,8 +1,9 @@
 use crate::{
     Error, Log,
     cdk::structures::{DefaultMemoryImpl, cell::Cell, memory::VirtualMemory},
-    icu_register_memory, impl_storable_bounded, log,
-    memory::{APP_STATE_MEMORY_ID, MemoryError},
+    icu_memory, impl_storable_bounded, log,
+    memory::{MemoryError, id::state::APP_STATE_ID},
+    thread_local_memory,
 };
 use candid::CandidType;
 use derive_more::Display;
@@ -10,11 +11,11 @@ use serde::{Deserialize, Serialize};
 use std::cell::RefCell;
 use thiserror::Error as ThisError;
 
-// thread local
-thread_local! {
+// APP_STATE
+thread_local_memory! {
     static APP_STATE: RefCell<Cell<AppStateData, VirtualMemory<DefaultMemoryImpl>>> =
         RefCell::new(Cell::init(
-            icu_register_memory!(APP_STATE_MEMORY_ID),
+            icu_memory!(AppState, APP_STATE_ID),
             AppStateData::default(),
         ));
 }